+++ examples/basic/monorise.config.ts
@@ -1,0 +1,7 @@
+const config = {
+  configDir: './src/monorise/entities',
+  // custom route file should export default an Hono object.
+  // customRoutes: './path/to/custom/routes.ts'
+};
+
+export default config;