+++ packages/cli/commands/build.ts
@@ -1,0 +1,8 @@
+import generateFiles from './utils/generate';
+
+async function runBuildCommand() {
+  console.log('Starting Monorise build...');
+  await generateFiles();
+}
+
+export default runBuildCommand;